Flask/requirements.txt
Flask==2.1.0
psycopg2-binary==2.9.5
pandas==1.5.3
scikit-learn==1.2.0
spacy==3.5.0